requirements.txt
aniso8601==1.3.0
appdirs==1.4.3
asn1crypto==0.22.0
boto==2.46.1
boto3==1.4.4
botocore==1.5.13
certifi==2017.11.5
cffi==1.9.1
chardet==3.0.4
click==6.7
coreapi==2.1.1
coreschema==0.0.4
cryptography==1.7.2
decorator==4.1.2
Django==1.11
django-appconf==1.0.2
django-bootstrap3==8.2.2
django-braces==1.11.0
django-compressor==2.1
django-cors-headers==1.3.1
django-cors-middleware==1.3.1
django-oauth-toolkit==0.12.0
django-rest-swagger==2.1.0
django-reversion==2.0.10
django-sslserver==0.19
django-storages==1.5.2
django-tracking2==0.4.0
django-widget-tweaks==1.4.1
djangorestframework==3.6.2
docutils==0.13.1
drf-tracking==1.2.0
dropbox==7.2.1
enum34==1.1.6
Fabric==1.13.1
functools32==3.2.3.post2
future==0.16.0
futures==3.0.5
google-api-python-client==1.6.2
httplib2==0.10.3
idna==2.6
ipaddress==1.0.18
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.9.6
jmespath==0.9.1
jsonpickle==0.9.4
jsonschema==2.6.0
jws==0.1.3
language-tags==0.4.3
lxml==4.1.1
MarkupSafe==1.0
MySQL-python==1.2.5
oauth2client==4.0.0
oauthlib==2.0.1
olefile==0.44
onedrivesdk==1.1.8
openapi-codec==1.2.1
openbadges==1.0.1
openbadges-bakery==1.0.0b4
packaging==16.8
paramiko==2.1.2
Pillow==4.0.0
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
pycrypto==2.6.1
pydux==0.2.2
PyJWT==1.4.2
PyLD==0.7.1
pyOpenSSL==16.2.0
pyparsing==2.2.0
PyPDF2==1.26.0
pypng==0.0.18
python-dateutil==2.6.0
python-magic==0.4.13
pytz==2017.2
rcssmin==1.0.6
reportlab==3.3.0
requests==2.18.4
requests-cache==0.4.13
rfc3986==0.4.1
rjsmin==1.0.12
rsa==3.4.2
s3transfer==0.1.10
simplejson==3.10.0
six==1.10.0
svglib==0.8.1
twilio==5.7.0
typing==3.6.1
uritemplate==3.0.0
urllib3==1.22
validators==0.11.2
webservices==0.7
